Dorada migracija - Ubacivanje migracije za promenu postojece kolone i dodavanje spoljnih kljuceva

+++ prodavnica_digitalnih_proizvoda/database/migrations/2024_12_24_104734_create_pictures_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('low_res_path')->default("images/low_res/defaultl.png");
             $table->string('high_res_path')->default("images/high_res/default.png");
             $table->decimal('price', 10, 2);
-            $table->unsignedBigInteger('downloads')->default(0);
+            $table->BigInteger('downloads')->default(0);
             $table->json('tags')->nullable();
         });
     }
